Store cowHerd health as double

+++ src/main/java/edu/ucsb/cs156/happiercows/entities/CowHerd.java
@@ -33,5 +33,5 @@ public class CowHerd {
     private int numCows;
 
     @Builder.Default
-    private int health = 100;
+    private double health = 100.0;
 }
